rename: 도메인 패키지 상세 분리
