+++ src/main/java/blackjack/domain/card/Cards.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -9,7 +9,6 @@ public class Cards {
     private static final int MAKE_ACE_BIGGER_SCORE = 10;
     private static final int BLACKJACK_SIZE_CONDITION = 2;
     private static final int BLACKJACK_SCORE_CONDITION = 21;
-    private static final int SPECIFIC_SCORE_OF_DEALER = 16;
 
     private final List<Card> cards;
