+++ src/main/java/blackjack/domain/participant/PlayersFactory.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import java.util.Arrays;
 import java.util.List;
