+++ src/test/java/blackjack/domain/card/DeckTest.java
@@ -1,6 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
