+++ src/test/java/blackjack/domain/participant/PlayersFactoryTest.java
@@ -1,9 +1,11 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.junit.jupiter.api.Assertions.*;
 
+import blackjack.domain.participant.Players;
+import blackjack.domain.participant.PlayersFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
